Cria modulo para ferramentas do parser
