+++ src/main/kotlin/compiler/Parser.kt
@@ -2,97 +2,8 @@ package compiler.parser
 
 import compiler.symbolTable.*
 import compiler.token.*
-import compiler.tokenMatch.*
+import compiler.parserTools.*
 
-
-data class ParserState(
-    val tokens: List<Token>,
-    var lookahead: Int,
-    var error: Throwable?,
-) {
-    fun next(): Token {
-        return tokens[lookahead]
-    }
-
-    fun incNew(): ParserState {
-        return ParserState(tokens, lookahead + 1, error)
-    }
-
-    fun errorNew(error_new: Throwable): ParserState {
-        return ParserState(tokens, lookahead, error_new)
-    }
-
-    fun hasToken(): Boolean {
-        return lookahead < tokens.size 
-    }
-
-    fun success(): Boolean {
-        return error == null
-    }
-}
-
-interface MatchParser {
-    fun match(state: ParserState): ParserState
-}
-
-class MatchString(val str: String): MatchParser {
-    override fun match(state: ParserState): ParserState {
-        if (!state.hasToken()) {
-            return state.errorNew(Throwable("É esperado o token $str, no entanto, o arquivo finalizou"))
-        }
-        val token = state.next()
-        if (token.tokenStr == str) {
-            return state.incNew()
-        } else {
-            return state.errorNew(Throwable("Na linha ${token.getLineNumber()} é esperado o token $str, ao invés de ${token.tokenStr}."))
-        }
-    }
-}
-
-class MatchStrings(val strs: List<String>): MatchParser {
-    override fun match(state: ParserState): ParserState {
-        var match_state = state
-        for (str in strs) {
-            val match_str = MatchString(str)
-            match_state = match_str.match(state)
-            if (match_state.success()) {
-                return match_state
-            }
-        }
-        return match_state
-    }
-}
-
-class MatchType(val type: TokenType): MatchParser {
-    override fun match(state: ParserState): ParserState {
-        if (!state.hasToken()) {
-            return state.errorNew(Throwable("É esperado token do tipo $type, no entanto, o arquivo finalizou"))
-        }
-        val token = state.next()
-        if (token.type == type) {
-            return state.incNew()
-        } else {
-            return state.errorNew(Throwable("Na linha ${token.getLineNumber()} é esperado token do tipo $type, ao invés de ${token.tokenStr}."))
-        }
-    }
-}
-
-class MatchFun(val func: (ParserState) -> ParserState): MatchParser {
-    override fun match(state: ParserState): ParserState {
-        return func(state)
-    }
-}
-
-fun checkMatches(matches: List<MatchParser>, state: ParserState): ParserState {
-    var s = state
-    for (match in matches) {
-        s = match.match(s)
-        if (!s.success()) {
-            return s
-        }
-    }
-    return s
-}
 
 fun run(tokens: List<Token>, table: SymbolTable): Result<Int> {
     val useful_tokens = tokens.filter { it.type != TokenType.ESPACO && it.type != TokenType.COMENTARIO }
@@ -107,6 +18,8 @@ fun run(tokens: List<Token>, table: SymbolTable): Result<Int> {
     }
 }
 
+// DEFINIÇÃO DE PRODUÇÃO
+// PROG -> program IDENTIFICADORES; | program IDENT; LISTF
 fun prog(state: ParserState): ParserState {
     val only_prog = listOf(MatchString("program"), MatchType(TokenType.IDENTIFICADORES), MatchString(";"), MatchFun(::endOfFile))
     val only_prog_match = checkMatches(only_prog, state)
@@ -126,6 +39,7 @@ fun prog(state: ParserState): ParserState {
     }
 }
 
+// DEFINIÇÃO DE PRODUÇÃO
 // DEC -> TYPE IDENT = PALAVRAS
 fun dec(state: ParserState): ParserState {
     val matches = listOf(
@@ -137,6 +51,7 @@ fun dec(state: ParserState): ParserState {
     return checkMatches(matches, state)
 }
 
+// DEFINIÇÃO DE PRODUÇÃO
 // ARGS -> TYPE IDENT, ARGS | TYPE IDENT | LAMBDA
 fun args(state: ParserState): ParserState {
     val list_args = listOf(MatchFun(::type), MatchType(TokenType.IDENTIFICADORES), MatchString(","), MatchFun(::args))
@@ -152,6 +67,8 @@ fun args(state: ParserState): ParserState {
     return state
 }
 
+// DEFINIÇÃO DE PRODUÇÃO
+// TYPE -> CHAR | INT | STRING | RATIONAL | FLOAT
 fun type(state: ParserState): ParserState {
     val match = MatchStrings(listOf("char", "int", "string", "rational", "float"))
     val match_state = checkMatches(listOf(match), state)
