+++ test/empty_program.uwu
@@ -1,0 +1,1 @@
+program hello;
