+++ test/program.uwu
@@ -1,0 +1,3 @@
+program hello;
+
+float nn, rational yy
